% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_det_banana.R
\name{banana}
\alias{banana}
\title{Rosenbrock's Banana Function}
\usage{
banana(x, scale01 = TRUE, ab = c(1, 100))
}
\arguments{
\item{x}{Inputs of dimension (at least) 2. See below for details.}

\item{scale01}{(No effect) When TRUE, inputs are expected to be given on unit interval.}

\item{ab}{Hyperparameters to Rosenbrocks banana}
}
\description{
Rosenbrock's Banana Function
}
\details{
A non-convec function introduced by Howard H. Rosenbrock in 1960. It is a difficult optimization test-problem, with a unique minimum at (1,1).
}
\examples{
fname <- "banana"
n <- 10
p <- quack(fname, verbose=FALSE)$input_dim
x <- matrix(runif(n*p), nrow=n)
y <- eval_duq(fname, x)
}
\references{
Rosenbrock, HoHo. "An automatic method for finding the greatest or least value of a function." The computer journal 3.3 (1960): 175-184.
}
