% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beam_deflection.R
\name{beam_deflection}
\alias{beam_deflection}
\title{Beam Deflection}
\usage{
beam_deflection(x, scale01 = TRUE, p = seq(0, 1, length.out = 20))
}
\arguments{
\item{x}{Inputs: Uniform load density (P, N/m), Elastic Modulus (E, Mpa), beam length (L, m), beam width (w, m), beam thickness (h, m)}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{p}{Vector of numbers between 0 and 1. p*L is location from the origin (left most side of beam) at which deflection is measured.}
}
\value{
The "output" vector is given by \code{-P/(24EI)ell(ell^3 - 2*L*ell^2 + L^3)} where \code{I = w*h^3/12} and \code{ell = p*L}
}
\description{
Dimensions 5. Measures the deflection of a length-L beam with a uniform load at location t away from origin.
}
\details{
A modified version of the beam deflection problem from the intro of Joseph (2024)
}
\examples{
n <- 10
x <- matrix(runif(3*n), nrow=n)
y <- apply(x, 1, beam_deflection)
}
\references{
Joseph, Roshan. "Rational Kriging". JASA, (2024)
}
