% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borehole.R
\name{borehole}
\alias{borehole}
\title{The Borehole Function}
\usage{
borehole(x, scale01 = TRUE, adjust_fidelity = 0)
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{adjust_fidelity}{Default value of 0 corresponds to the usual borehole function. Value of 1 corresponds to the low-fidelity version described in the VLSE (Xiong 2013).}
}
\value{
Flow through a borehole.
}
\description{
Dimensions 8. The Borehole function models water flow through a borehole. Its simplicity and quick evaluation makes it a commonly used function for testing a wide variety of methods in computer experiments.
}
\details{
For details on the borehole function, see \href{https://www.sfu.ca/~ssurjano/borehole.html}{the VLSE}.
Parameter ranges:
\describe{
  \item{x1 = rw}{radius of borehole (m), [0.05, 0.15]}
  \item{x2 = r}{radius of influence (m), [100, 50000]}
  \item{x3 = Tu}{transmissivity of borehole (m^2/yr), [63070, 115600]}
  \item{x4 = Hu}{potentiometric head of upper aquifer (m), [990, 1110]}
  \item{x5 = Tl}{transmissivity of lower aquifer (m2/yr), [63.1, 116]}
  \item{x6 = Hl}{potentiometric head of lower aquifer (m), [700, 820]}
  \item{x7 = L}{length of borehole (m), [1120, 1680]}
  \item{x8 = Kw}{hydraulic conductivity of borehole (m/yr), [9855, 12045]}
}
For a more non-linear and non-additive function (Xiong 2013), the range of Kw may be changed to [1500, 15 000]. Fidelity can be adjusted to match the low-fidelity borehole function of Xiong 2013.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, borehole, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Harper, W. V., & Gupta, S. K. (1983). Sensitivity/uncertainty analysis of a borehole scenario comparing Latin Hypercube Sampling and deterministic sensitivity approaches (No. BMI/ONWI-516). Battelle Memorial Inst., Columbus, OH (USA). Office of Nuclear Waste Isolation.

Xiong, S., Qian, P. Z., & Wu, C. J. (2013). Sequential design and analysis of high-accuracy and low-accuracy computer codes. Technometrics, 55(1), 37-46.
}
