% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_funcs.R
\name{borehole_cat}
\alias{borehole_cat}
\title{Mixed Inputs Borehole Function}
\usage{
borehole_cat(x, z, scale01 = TRUE, log = TRUE, case = 1)
}
\arguments{
\item{x}{Quantitative inputs of dimension 2.}

\item{z}{Binary-coded categorical inputs of dimension 1}

\item{scale01}{When \code{TRUE}, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{log}{Should the response be log-transformed (as in Zhang et al (2021)).}

\item{case}{Corresponds to the cases in Table 2 of Zhang et al (2021). The number of qualitative levels is equal to case1: 27, case2: 40, case3: 60, case4: 80.}
}
\value{
A scalar response.
}
\description{
Dimensions: 4 continuous, 3 categorical. A test function with continuous and categorical inputs.
}
\details{
For details, see \href{https://www.sfu.ca/~ssurjano/moon10mix.html}{the VLSE}.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, pollutant, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Zhou, Q., Qian, P. Z., & Zhou, S. (2011). A simple approach to emulation for computer models with qualitative and quantitative factors. Technometrics, 53(3).

Zhang, Q., Chien, P., Liu, Q., Xu, L., & Hong, Y. (2021). Mixed-input Gaussian process emulators for computer experiments with a large number of categorical levels. Journal of Quality Technology, 53(4), 410-420.
}
