% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borehole_low_fidelity.R
\name{borehole_low_fidelity}
\alias{borehole_low_fidelity}
\title{The Low Fidelity Borehole Function}
\usage{
borehole_low_fidelity(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
Flow through a borehole.
}
\description{
Dimensions 8. Low fidelity version of the borehole function (Xiong 2013). See \code{\link[duqling]{borehole}} for more details.
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Xiong, S., Qian, P. Z., & Wu, C. J. (2013). Sequential design and analysis of high-accuracy and low-accuracy computer codes. Technometrics, 55(1), 37-46.
}
