% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackscholes.R
\name{bs_call}
\alias{bs_call}
\alias{bs_put}
\title{Stochastic Black-Scholes formula}
\usage{
bs_call(x, scale01 = TRUE, time = seq(from = 0, to = 1, length.out = 30))

bs_put(x, scale01 = TRUE, time = seq(from = 0, to = 1, length.out = 30))
}
\arguments{
\item{x}{Inputs of dimension (at least) 6. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{time}{a vector of times (on a (0, 1) scale).}
}
\value{
Put and call prices for an asset.
}
\description{
Dimensions 6. Models the dynamics of a financial market (put and call prices)
under a stochastic model for the underlying asset.
}
\details{
A stochastic version of the Black-Scholes model.
\describe{
  \item{x1 = sigma}{Volatility of the market, \eqn{\sigma \in [0.02, 0.6]}.}
  \item{x2 = T}{Time to maturity (years), \eqn{T \in [0.25, 2.0]}.}
  \item{x3 = K}{Strike price, \eqn{K \in [50, 150]}.}
  \item{x4 = r}{Risk-free interest rate, \eqn{r \in [0, 0.1]}.}
  \item{x5 = S0}{Initial underlying asset price, \eqn{S_0 \in [80, 120]}.}
  \item{x6 = mu}{Drift of the underlying asset under the real-world measure, \eqn{\mu \in [-0.1, 0.2]}.}
}
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, bs_, scale01=TRUE)
}
\references{
Black, Fischer (1976). The pricing of commodity contracts, Journal of Financial Economics, 3, 167-179.
}
