% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_funcs.R
\name{cat_map}
\alias{cat_map}
\title{Here we collect some test functions with categorical inputs.
For ease of use and consistency with the systematic behavior of the
future function: run_sim_study_cat(), we re-code all categorical variables
using binary coding. The metadata fields $categorical_map and $num_classes
are enough to do this on the original scale, if desired.}
\usage{
cat_map(bits, num_classes, clamp = TRUE)
}
\description{
Here we collect some test functions with categorical inputs.
For ease of use and consistency with the systematic behavior of the
future function: run_sim_study_cat(), we re-code all categorical variables
using binary coding. The metadata fields $categorical_map and $num_classes
are enough to do this on the original scale, if desired.
}
