% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otl_circuit.R
\name{circuit}
\alias{circuit}
\title{OTL Circuit Function}
\usage{
circuit(x, scale01 = FALSE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 7. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
The response Vm is midpoint voltage
}
\description{
Dimensions: 6 The OTL Circuit function models an output transformerless push-pull circuit.
}
\details{
For details on this function, see \href{https://www.sfu.ca/~ssurjano/otlcircuit.html}{the VLSE}.
Parameter ranges:
\describe{
  \item{x1 = Rb1}{resistance b1 (K-Ohms), [50, 150]}
  \item{x2 = Rb2}{resistance b2 (K-Ohms), [25, 70]}
  \item{x3 = Rf}{resistance f (K-Ohms),   [0.5, 3]}
  \item{x4 = Rc1}{resistance c1 (K-Ohms), [1.2, 2.5]}
  \item{x5 = Rc2}{resistance c2 (K-Ohms), [0.25, 1.2]}
  \item{x5 = beta}{current gain (Amps), [50, 300]}
}
}
\examples{
n <- 10
x <- matrix(runif(6*n), nrow=n)
y <- apply(x, 1, circuit, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Ben-Ari, E. N., & Steinberg, D. M. (2007). Modeling data from computer experiments: an empirical comparison of kriging with MARS and projection pursuit regression. Quality Engineering, 19(4), 327-338.
}
