% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_det_constant.R
\name{constant}
\alias{constant}
\alias{const_fn}
\alias{const_fn3}
\alias{const_fn15}
\title{Constant Functions}
\usage{
const_fn(x, scale01 = TRUE)

const_fn3(x, scale01 = TRUE)

const_fn15(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of any dimension. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\description{
Constant Functions
}
\details{
Always returns 0
}
\examples{
fname <- "const_fn3"
n <- 10
p <- quack(fname, verbose=FALSE)$input_dim
x <- matrix(runif(n*p), nrow=n)
y <- eval_duq(fname, x)

}
