% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{cubes5}
\alias{cubes5}
\alias{cube5}
\title{Discontinuous Cube Functions}
\usage{
cube5(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{(No effect here) When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\description{
Discontinuous Cube Functions
}
