% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehundred.R
\name{d_onehundred}
\alias{d_onehundred}
\title{The Gradient of the 100D Function}
\usage{
d_onehundred(x, scale01 = TRUE, M = 100)
}
\arguments{
\item{x}{Inputs of dimension (at least) 2. See below for details.}

\item{scale01}{(No effect) When TRUE, inputs are expected to be given on unit interval.}

\item{M}{The number of active variables (default M=100, (54 <= M <= 100))}
}
\description{
Input Dimension: 100, Response Dimension: 100
}
\details{
The 100-dimensional function is a high-dimensional function derived from a more generic example detailed below. This function was designed for sensitivity analysis: the first-order sensitivity indices of the input variables generally nonlinearly increase with their index, with certain variables having especially high sensitivity (Lüthen et al, 2021).
}
\examples{
n <- 1000
p <- 100
X <- matrix(runif(n*p), nrow=n, ncol=p)
y <- apply(X, 1, d_onehundred, scale01=TRUE)
# not run
# fit <- basssPCA(X, t(y))
}
\references{
Lüthen, Nora, Stefano Marelli, and Bruno Sudret. "Sparse polynomial chaos expansions: Literature survey and benchmark." SIAM/ASA Journal on Uncertainty Quantification 9.2 (2021): 593-649.

UQLab - The Framework for Uncertainty Quantification. “Sensitivity: Sobol’ indices of a high-dimensional function.” Retrieved July 3, 2024, from https://www.uqlab.com/sensitivity-high-dimension.
}
