% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R, R/dms.R
\name{cube3}
\alias{cube3}
\alias{cube3_rotate}
\alias{cube5}
\alias{dms}
\alias{dms_simple}
\alias{dms_radial}
\alias{dms_harmonic}
\alias{dms_additive}
\alias{dms_complicated}
\title{Denison, Mallick and Smith Bivariate Test Functions}
\usage{
cube3(x, scale01 = TRUE)

cube3_rotate(x, scale01 = TRUE)

cube5(x, scale01 = FALSE)

dms_simple(x, scale01 = FALSE)

dms_radial(x, scale01 = FALSE)

dms_harmonic(x, scale01 = FALSE)

dms_additive(x, scale01 = FALSE)

dms_complicated(x, scale01 = FALSE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{(No effect here) When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\description{
Denison, Mallick and Smith Bivariate Test Functions
}
\details{
The five test functions from Denison, Mallick and Smith (1998) are referred to as
\enumerate{
   \item Simple
   \item Radial
   \item Harmonic
   \item Additive
   \item Complex interaction
}
}
\examples{
n <- 100 #Number of observations
p <- 3   #Number of variables
X <- matrix(runif(n*p), nrow=n)
y <- apply(X, 1, cube3)
n <- 100 #Number of observations
p <- 4   #Number of variables (beyond p = 2, variables are inert)
X <- matrix(runif(n*p), nrow=n)
y <- apply(X, 1, dms_simple)
}
\references{
Denison, David GT, Bani K. Mallick, and Adrian FM Smith. "Bayesian mars." Statistics and Computing 8.4 (1998): 337-346.
}
