% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duq.R
\name{duq}
\alias{duq}
\title{duq}
\usage{
duq(x, f, scale01 = TRUE, ...)
}
\arguments{
\item{x}{An nxp matrix of inputs.}

\item{f}{A string (of a function name) or a function, usually from the duqling package}

\item{scale01}{When TRUE, inputs are expected on the (0, 1) scale.}

\item{...}{additional functions to be passed to f}
}
\value{
See details
}
\description{
Call functions from duqling namespace
}
\details{
A shortcut for calling funtions in the duqling package
}
\examples{
n <- 100
p <- 10
X <- lhs::randomLHS(n, p)
duq(X, borehole)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.
}
