% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duqling-package.R
\docType{package}
\name{duqling-package}
\alias{duqling}
\alias{duqling-package}
\title{duqling}
\description{
An R package containing a large number of popular UQ test functions.
These functions facilitate reproducible UQ research.
No more looking up function definition or input ranges ever again!

  Every duqling function
has a the \code{scale01} argument which, when TRUE, lets the user pass inputs on a (0, 1)
range and the inputs are scaled internally.

The \code{run_sim_study()} function makes running
simulation studies easy. The comparison of various emulators is also straightforward, even
across papers.
}
\keyword{internal}
