% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola.R
\name{ebola}
\alias{ebola}
\title{The Ebola Function}
\usage{
ebola(x, scale01 = TRUE, location = "liberia")
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range.}

\item{location}{Default value is "liberia" but "sierra_leone" is also accepted.}
}
\value{
Basic reproduction number
}
\description{
Dimensions 8. Calculates the estimated basic reproduction number, R0, for an Ebola outbreak.
}
\details{
See section 5.1 of reference for details and further references.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, ebola)
}
\references{
Duan, Hui, and Giray Okten. "Derivative-based Shapley value for global sensitivity analysis and machine learning explainability." International Journal for Uncertainty Quantification 15.1 (2025).
}
