% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duq.R
\name{eval_duq}
\alias{eval_duq}
\title{Evaluate a UQ Test Function}
\usage{
eval_duq(f, x, z = NULL, scale01 = TRUE, ...)
}
\arguments{
\item{f}{A string (of a function name) or a function, usually from the duqling package}

\item{x}{A matrix of categorical inputs with \code{n} rows and \code{quack()$input_dim} columns.}

\item{z}{A matrix of categorical inputs with \code{n} rows and \code{quack()$input_cat_dim} columns. Categories should be coercible to an integer between 1 and number of levels.}

\item{scale01}{When \code{TRUE}, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{...}{additional functions to be passed to f}
}
\value{
See details
}
\description{
Call functions from duqling namespace
}
\details{
A shortcut for calling funtions in the duqling package
}
\examples{
n <- 100
p <- 2
X <- lhs::randomLHS(n, p)
f_list <- quack(input_dim=2)$fname
eval_duq(f_list[1], X)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.
}
