% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forrester1.R
\name{forrester1}
\alias{forrester1}
\title{1D Multifidelity Function}
\usage{
forrester1(x, scale01 = TRUE, A = 1, B = 0, C = 0)
}
\arguments{
\item{x}{Inputs of dimension (at least) 1.}

\item{scale01}{Ignored.}

\item{A}{Parameter for low fidelity version (see details)}

\item{B}{Parameter for low fidelity version (see details)}

\item{C}{Parameter for low fidelity version (see details)}
}
\value{
Scalar output
}
\description{
Dimensions: 1. This function is a simple one-dimensional test function. It is multimodal, with one global minimum, one local minimum and a zero-gradient inflection point.
}
\details{
A simple 1D test function. Low fidelity version is simulated as Af(x) + B(x-1/2) - C. See reference for details.
}
\examples{
n <- 1
x <- matrix(runif(20), nrow=n)
y <- apply(x, 1, forrester1)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Forrester, A., Sobester, A., & Keane, A. (2008). Engineering design via surrogate modelling: a practical guide. Wiley.
}
