% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foursquare.R
\name{foursquare}
\alias{foursquare}
\title{The Foursquare Function}
\usage{
foursquare(x, scale01 = TRUE, ftype = "all")
}
\arguments{
\item{x}{Inputs of dimension (at least) 7. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
The response C is cycle time (the time it takes to complete one cycle), in seconds
}
\description{
Dimensions: 2
}
\details{
Each of the four quadrants can be modeled effectively with a different approach (rbf, linear, mars, trees).
The function response is $y = f1 + f2 + f3 + f4$ where
\describe{
  \item{f1}{An RBF centered at (0.75, 0.75) with variance of 0.13 and a correlation of -0.5}
  \item{f2}{The linear plane x_2 - 1.375*x_1}
  \item{f3}{A mars basis function with a knot at (0.35, 0.35)}
  \item{f4}{A scaled/shifted version of \code{duqling::sharkfin}, designed for tree based methods.}
}
}
\examples{
n <- 10
p <- 2
x <- matrix(runif(p*n), nrow=n)
y <- apply(x, 1, foursquare, scale01=TRUE)
}
