% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedman.R
\name{friedman}
\alias{friedman}
\alias{friedman10}
\alias{friedman20}
\title{The Friedman Function}
\usage{
friedman(x, scale01 = TRUE)

friedman10(x, scale01 = TRUE)

friedman20(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 5. See below for details.}

\item{scale01}{(Ignored here) When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
A scalar response
}
\description{
Dimensions: 5. The Friedman function is often used with additional, inert inputs.
}
\details{
For details on the Friedman function, see \href{https://www.sfu.ca/~ssurjano/fried.html}{the VLSE}.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, friedman, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Friedman, J. H. (1991). Multivariate adaptive regression splines. The annals of statistics, 19(1), 1-67.
}
