% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_det_gamma_mix.R
\name{gamma_mix}
\alias{gamma_mix}
\title{The Gamma Mix Function}
\usage{
gamma_mix(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 7. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be on the unit interval and are internally adjusted to their native range.}
}
\value{
Value of the gamma mixture density at the location x[1].
}
\description{
Dimensions: 7. Returns the value of a two-component gamma mixture density at a given location, with user-defined parameters and an optional horizontal shift.
}
\details{
The function is defined as:
\deqn{f(x) = w \cdot \text{dgamma}(x_1 - m; a_1, b_1) + (1-w) \cdot \text{dgamma}(x_1 - m; a_2, b_2)}
where
\itemize{
  \item \code{x[1]}: location to evaluate (default range [0, 50])
  \item \code{x[2]}: shape parameter \code{a1} for first gamma ([1, 10])
  \item \code{x[3]}: rate parameter \code{b1} for first gamma ([0.5, 5])
  \item \code{x[4]}: shape parameter \code{a2} for second gamma ([1, 10])
  \item \code{x[5]}: rate parameter \code{b2} for second gamma ([0.5, 5])
  \item \code{x[6]}: mixing weight \code{w} ([0, 1])
  \item \code{x[7]}: horizontal shift \code{m} ([0, 10])
}
The density is evaluated at \code{x[1]} minus the shift \code{m}. If the argument to dgamma is negative, the density is zero.
}
\examples{
fname <- "gamma_mix"
n <- 10
p <- quack(fname, verbose=FALSE)$input_dim
x <- matrix(runif(n*p), nrow=n)
y <- eval_duq(fname, x)
}
