% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_UQ_data.R
\name{get_emulation_data}
\alias{get_emulation_data}
\title{Get UQ data for emulation}
\usage{
get_emulation_data(dname)
}
\arguments{
\item{name}{String. Name of dataset (see documentation)}
}
\value{
A list with elements `y` (a univariate response) and `X` (a matrix of inputs).
}
\description{
Load in UQ datasets from the UQdataverse (harvard)
}
\details{
Wrapper to read in various UQ datasets (from harvard dataverse)
#' @details Wrapper to read in various UQ datasets (from harvard dataverse)
\itemize{
 \item{"strontium_plume_"}{Input dimension = 20, observations = 300}
 \item{"parameter 2"}{Stuff}
}
}
\examples{
get_emulation_data
}
