% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.R
\name{get_funcs}
\alias{get_funcs}
\alias{get_sim_functions_full}
\alias{get_sim_functions_2d}
\alias{get_sim_functions_medium}
\alias{get_sim_functions_short}
\alias{get_sim_functions_tiny}
\title{Sim Study Function Lists}
\usage{
get_sim_functions_full()

get_sim_functions_2d()

get_sim_functions_medium()

get_sim_functions_short()

get_sim_functions_tiny()
}
\description{
Sim Study Function Lists
}
