% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_names.R
\name{get_sim_data}
\alias{get_sim_data}
\alias{get_sim_data_full}
\alias{get_sim_data_tiny}
\alias{get_sim_data_medium}
\alias{get_sim_data_2d}
\title{Sim Study Datset Lists}
\usage{
get_sim_data_full()

get_sim_data_tiny()

get_sim_data_medium()

get_sim_data_2d()
}
\description{
Sim Study Datset Lists
}
