% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_UQdata.R
\name{get_UQdata}
\alias{get_UQdata}
\title{Get UQ data}
\usage{
get_UQdata(dname, control = list())
}
\arguments{
\item{control}{A list of control variables (currently not used)}

\item{name}{String. Name of dataset (see documentation)}
}
\value{
Data frame
}
\description{
Load in UQ datasets from the UQdataverse (harvard)
}
\details{
Wrapper to read in various UQ datasets (from harvard dataverse)
\itemize{
 \item{"strontium_plume"}{Input dimension = 20, observations = 300}
 \item{"parameter 2"}{Stuff}
}
}
\examples{
data <- get_UQdata("strontium_plume")
}
