% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gfunctions.R
\name{Gfunction}
\alias{Gfunction}
\alias{Gfunction6}
\alias{Gfunction12}
\alias{Gfunction18}
\title{The Gfunction Function}
\usage{
Gfunction6(x, scale01 = TRUE, a = c(0, 0, rep(6.52, 4)))

Gfunction12(x, scale01 = TRUE, a = c(0, 0, rep(6.52, 4), 9, 9, 15, 25, 50, 99))

Gfunction18(
  x,
  scale01 = TRUE,
  a = c(rep(0, 4), rep(1, 4), rep(9, 4), rep(18, 4), rep(99, 2))
)
}
\arguments{
\item{x}{Inputs of dimension (at least) 5. See below for details.}

\item{scale01}{(Ignored here) When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
A scalar response
}
\description{
Dimensions: d. A multiplicative function.
}
\details{
The G function function is often used as an integrand for various numerical estimation methods. The .exact value of the intregal of this function in any dimensions is 1. The a_i values indicate the importance of a variable (lower value is more important).
For details on the G function function, see \href{https://www.sfu.ca/~ssurjano/fried.html}{the VLSE}.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, Gfunction, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Gfunction, J. H. (1991). Multivariate adaptive regression splines. The annals of statistics, 19(1), 1-67.
}
