% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_det_higdon1.R
\name{higdon1}
\alias{higdon1}
\title{1D Higdon Function}
\usage{
higdon1(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 1.}

\item{scale01}{Are inputs expected on (0, 1) scale?.}
}
\value{
Scalar output
}
\description{
Dimensions: 1. This function is a simple one-dimensional test function exhibiting non-stationarity.
}
\details{
A simple 1D test function. Low fidelity version is simulated as Af(x) + B(x-1/2) - C. See reference for details.
}
\examples{
fname <- "higdon1"
n <- 10
p <- quack(fname, verbose=FALSE)$input_dim
x <- matrix(runif(n*p), nrow=n)
y <- eval_duq(fname, x)
}
\references{
Gramacy, R.B., Lee, H.K. and Macready, W.G., 2004, July. Parameter space exploration with Gaussian process trees. In Proceedings of the twenty-first international conference on Machine learning (p. 45)
}
