% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_ignition.R
\name{ignition}
\alias{ignition}
\title{Mock Ingition Function}
\usage{
ignition(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 10. See below for details.}

\item{scale01}{(No effect here). When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range.}
}
\value{
The response C is cycle time (the time it takes to complete one cycle), in seconds
}
\description{
Dimensions: 10 The complex mock ignition function from Section E of Hatfield et al. 2019.
}
\details{
From Hatfield et al. 2019, the mock-yield is r^5(1 + 100000 + (1+erf(10(r-2)))) where r is the radius from the origin in this 10-D space.
}
\examples{
p <- 10
n <- 20
x <- matrix(runif(n*p), nrow=n)
y <- apply(x, 1, ignition, scale01=TRUE)
}
\references{
Hatfield, P., Rose, S., Scott, R., Almosallam, I., Roberts, S., & Jarvis, M. (2019). Using sparse Gaussian processes for predicting robust inertial confinement fusion implosion yields. IEEE Transactions on Plasma Science, 48(1), 14-21.
}
