% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_det_linear.R
\name{linear}
\alias{linear}
\alias{linear21}
\alias{linear21_s10}
\alias{linear21_s1}
\alias{quad}
\alias{quad4}
\title{Linear and Quadratic Test Functions}
\usage{
linear(x, scale01 = TRUE, a = NULL)

linear21(x, scale01 = TRUE)

linear21_s10(x, scale01 = TRUE)

linear21_s1(x, scale01 = TRUE)

quad(x, scale01 = TRUE, a = NULL, b = NULL, c = NULL)

quad4(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension at least p.}

\item{scale01}{Are inputs given on (0, 1) scale?}

\item{a}{Vector of coefficients for linear terms.}

\item{b}{Vector of coefficients for quadratic terms.}

\item{c}{Vector of coefficients for pairwise interaction terms.}
}
\value{
Scalar output.
}
\description{
Simple linear and quadratic benchmark functions.
}
\details{
Simple linear or quadratic models.
}
\examples{
fname <- "linear21"
n <- 10
p <- quack(fname, verbose=FALSE)$input_dim
x <- matrix(runif(n*p), nrow=n)
y <- eval_duq(fname, x)
}
