% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_cat_funcs.R
\name{moon3_cat}
\alias{moon3_cat}
\title{Moon (2010) mixed function}
\usage{
moon3_cat(x, z, scale01 = TRUE)
}
\arguments{
\item{x}{Quantitative inputs of dimension 2.}

\item{z}{Binary-coded categorical inputs of dimension 1}

\item{scale01}{When \code{TRUE}, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
A scalar response.
}
\description{
Dimensions 2 continuous, 1 categorical. A test function with continuous and categorical inputs, designed for sensitivity analysis and screening
}
\details{
For details, see \href{https://www.sfu.ca/~ssurjano/moon10mix.html}{the VLSE}.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, pollutant, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Moon, H. (2010). Design and Analysis of Computer Experiments for Screening Input Variables (Doctoral dissertation, Ohio State University).
}
