% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micwicz.R
\name{multivalley}
\alias{multivalley}
\title{The Multivalley Function}
\usage{
multivalley(x, scale01 = TRUE, m = 10, active_dim = length(x))
}
\arguments{
\item{x}{Inputs of dimension (at least) 2. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{m}{A larger m leads to a more difficult search (steeper valleys/ridges).}

\item{active_dim}{Only the first \code{min(active_dim, length(x))} variables will be used. All other variables are inert.}
}
\value{
Function output
}
\description{
Dimension: Default is 2, but can be any dimension. For inert variables, \code{active_dim} must be specified.
}
\details{
Also called the Michalewicz function. For details on this function, see \href{https://www.sfu.ca/~ssurjano/michal.html}{the VLSE}.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, multivalley, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
}
