% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_circ.R
\name{ocean_circ}
\alias{ocean_circ}
\title{Stochastic Ocean Circulation Model}
\usage{
ocean_circ(x, scale01 = TRUE, NPATHS = 1)
}
\arguments{
\item{x}{Inputs of dimension (at least) 4. See below for details.}

\item{scale01}{Default TRUE. For this function inputs are expected on a [0,1] range.}

\item{NPATHS}{how many simulations should be conducted for each set of inputs (the average is returned).}
}
\value{
Let C(s) denote the concentration of dissolved oxygen at a location s = (lon, lat) in a rectangular domain Omega subset R2 representing the layer of the ocean being studied
}
\description{
Dimensions 4. Longitude, latitude, and the diffusion coefficients Kx and Ky
}
\details{
a stochastic simulator
that models the concentration of oxygen in a thin water
layer (around 2000 m deep) in the South Atlantic ocean
(McKeague et al., 2005, Herbei and Berliner, 2014).
For a specific
location within the domain, random paths of the process
are generated, producing noisy outcomes that approximate
the solution to the PDE at that location.
}
\examples{
n <- 10
x <- matrix(runif(4*n), nrow=n)
y <- apply(x, 1, ocean_circ)
}
\references{
Baker, Evan, et al. "Analyzing stochastic computer models: a review with opportunities." Statistical Science 37.1 (2022): 64-89.

Herbei, Radu, and L. Mark Berliner. "Estimating ocean circulation: an MCMC approach with approximated likelihoods via the Bernoulli factory." Journal of the American Statistical Association 109.507 (2014): 944-954.
}
