% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehundred.R
\name{onehundred}
\alias{onehundred}
\title{The 100D Function}
\usage{
onehundred(x, scale01 = TRUE, M = 100)
}
\arguments{
\item{x}{Inputs of dimension (at least) 2. See below for details.}

\item{scale01}{(No effect) When TRUE, inputs are expected to be given on unit interval.}

\item{M}{The number of active variables (default M=100, (54 <= M <= 100))}
}
\description{
Dimension: 100
}
\details{
The 100-dimensional function is a high-dimensional function derived from a more generic example detailed below. This function was designed for sensitivity analysis: the first-order sensitivity indices of the input variables generally nonlinearly increase with their index, with certain variables having especially high sensitivity (Lüthen et al, 2021).
}
\examples{
n1 <- 100 #Number of observations in x1
n2 <- 70  #Number of observations in x2
x1grid <- seq(0, 1, length.out=100)
x2grid <- seq(0, 1, length.out=70)
X <- expand.grid(x1grid, x2grid)
y <- apply(X, 1, banana, scale01=TRUE)
image(matrix(y, nrow=length(X)), zlim=c(-1,3.5))
}
\references{
Lüthen, Nora, Stefano Marelli, and Bruno Sudret. "Sparse polynomial chaos expansions: Literature survey and benchmark." SIAM/ASA Journal on Uncertainty Quantification 9.2 (2021): 593-649.

UQLab - The Framework for Uncertainty Quantification. “Sensitivity: Sobol’ indices of a high-dimensional function.” Retrieved July 3, 2024, from https://www.uqlab.com/sensitivity-high-dimension.
}
