% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park.R
\name{park}
\alias{park}
\alias{park4}
\alias{park4_low_fidelity}
\title{The Park Function}
\usage{
park4(x, scale01 = TRUE)

park4_low_fidelity(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 3. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval.}
}
\description{
Dimensions: 4.
}
\details{
Intended to be a simple four-dim test function. Xiong et al. (2013) use \code{park4_low_fidelity} as a low fidelity version. For details on the Friedman function, see \href{https://www.sfu.ca/~ssurjano/park91a.html}{the VLSE}.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, friedman, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Park, Jeong Soo. Tuning complex computer codes to data and optimal designs. University of Illinois at Urbana-Champaign, 1991.

Xiong, Shifeng, Peter ZG Qian, and CF Jeff Wu. "Sequential design and analysis of high-accuracy and low-accuracy computer codes." Technometrics 55.1 (2013): 37-46.
}
