% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permdb.R
\name{permdb}
\alias{permdb}
\title{The Permdb Function}
\usage{
permdb(x, scale01 = TRUE, d = 16, beta = 0.5)
}
\arguments{
\item{x}{Inputs of dimension (at least) \code{d}. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range.}

\item{d}{Dimension of the input (default 16).}

\item{beta}{Beta parameter in the perm function (default 0.5).}
}
\value{
Scalar output.
}
\description{
Dimensions: Variable (default 16). The permdb function is a well-known deterministic test function used in high-dimensional optimization and uncertainty quantification. It is challenging due to its intricate global structure and numerous local minima.
}
\details{
The permdb function is defined as:
\deqn{f(x) = \sum_{j=1}^d \left[ \sum_{i=1}^d (i^j + \beta) \left( \left(\frac{x_i}{i}\right)^j - 1 \right) \right]^2}
with the typical input domain \eqn{x_i \in [-d, d]}.
}
\examples{
n <- 3
x <- matrix(runif(n*16), nrow=n)
y <- apply(x, 1, permdb, scale01=TRUE)
}
\references{
Dixon, L. C. W., & Price, R. (1981). The perm function. In Towards Global Optimization 2 (pp. 169–204).
}
