% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piston.R
\name{piston}
\alias{piston}
\title{Piston Simulation Function}
\usage{
piston(x, scale01 = FALSE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 7. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
The response C is cycle time (the time it takes to complete one cycle), in seconds
}
\description{
Dimensions: 7 The Piston Simulation function models the circular motion of a piston within a cylinder. It involves a chain of nonlinear functions.
}
\details{
For details on the piston function, see \href{https://www.sfu.ca/~ssurjano/piston.html}{the VLSE}.
Parameter ranges:
\describe{
  \item{x1 = M}{piston weight (kg), [30, 60]}
  \item{x2 = S}{piston surface area (m^2), [0.005, 0.020]}
  \item{x3 = V0}{initial gas volume (m^3), [0.002, 0.010]}
  \item{x4 = k}{spring coefficient (N/m), [1000, 5000]}
  \item{x5 = P0}{atmospheric pressure (N/m^2), [90000, 110000]}
  \item{x6 = Ta}{ambient temperature (K), [290, 296]}
  \item{x7 = T0}{filling gas temperature (K), [340, 360]}
}
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, piston, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Kenett, R., & Zacks, S. (1998). Modern industrial statistics: design and control of quality and reliability. Pacific Grove, CA: Duxbury press.
}
