% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutant.R
\name{pollutant}
\alias{pollutant}
\title{The Environmental Model for Pollutant Spill in a Channel}
\usage{
pollutant(
  x,
  scale01 = TRUE,
  space = c(0.5, 1, 1.5, 2, 2.5),
  time = seq(from = 0.3, to = 60, by = 0.3)
)
}
\arguments{
\item{x}{Inputs of dimension (at least) 4. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{space}{a vector of locations along the channel (0 is location of the first spill).}

\item{time}{a vector of times. (0 is the time of the first spill)}
}
\value{
C(x) is the concentration of the pollutant at the space-time vector (s, t)
}
\description{
Dimensions 4. Output dimension is a space-time field. The Environmental Model function models a pollutant spill caused by a chemical accident. C(x) is the concentration of the pollutant at the space-time vector (s, t), where 0 ≤ s ≤ 3 and t > 0. I is the indicator function.
}
\details{
For details on the pollutant function, see \href{https://www.sfu.ca/~ssurjano/environ.html}{the VLSE}.
Parameter ranges:
\describe{
  \item{x1 = M}{mass of pollutant spilled at each location, [7, 13]}
  \item{x2 = D}{diffusion rate in the channel, [0.02, 0.12]}
  \item{x3 = L}{location of the second spill, [0.01, 3]}
  \item{x4 = tau}{time of the second spill, [30.01, 30.295]}
}
For a univariate version of this function, see \code{\link[duqling]{pollutant_uni}}
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, pollutant, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Bliznyuk, N., Ruppert, D., Shoemaker, C., Regis, R., Wild, S., & Mugunthan, P. (2008). Bayesian calibration and uncertainty analysis for computationally expensive models using optimization and radial basis function approximation. Journal of Computational and Graphical Statistics, 17(2).
}
