% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutant_uni.R
\name{pollutant_uni}
\alias{pollutant_uni}
\title{Univariate Pollutant Spill in a Channel}
\usage{
pollutant_uni(x, scale01 = TRUE, space = 2.5, time = 30)
}
\arguments{
\item{x}{Inputs of dimension (at least) 4. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{space}{single location along the channel (default 2.5)}

\item{time}{single time of measurement (default 30)}
}
\value{
C(x) is the concentration of the pollutant at the space-time vector (s, t)
}
\description{
Dimensions 4. The \code{\link[duqling]{pollutant}} function evaluated at time 30 and space = 2.5
}
\details{
For details on the pollutant function, see \href{https://www.sfu.ca/~ssurjano/environ.html}{the VLSE}.
Parameter ranges:
\describe{
  \item{x1 = M}{mass of pollutant spilled at each location, [7, 13]}
  \item{x2 = D}{diffusion rate in the channel, [0.02, 0.12]}
  \item{x3 = L}{location of the second spill, [0.01, 3]}
  \item{x4 = tau}{time of the second spill, [30.01, 30.295]}
}
For the spatio-temporal version of this model, see \code{\link[duqling]{pollutant}}
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, pollutant_uni, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Bliznyuk, N., Ruppert, D., Shoemaker, C., Regis, R., Wild, S., & Mugunthan, P. (2008). Bayesian calibration and uncertainty analysis for computationally expensive models using optimization and radial basis function approximation. Journal of Computational and Graphical Statistics, 17(2).
}
