% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quackquack.R
\name{quack}
\alias{quack}
\title{Quack (Gives Information about functions in this package)}
\usage{
quack(
  fname = NULL,
  input_dim = NULL,
  input_cat = NULL,
  response_type = NULL,
  stochastic = NULL,
  sorted = TRUE
)
}
\arguments{
\item{fname}{A string, the name of a function in this package}

\item{input_dim}{a vector specifying the input dimension of the function}

\item{input_cat}{logical, should functions with categorical inputs be included?}

\item{response_type}{a string in the set c("all", "uni", "multi", "func") specifying which response type is requested.}

\item{stochastic}{logical. Is function response stochastic?}

\item{sorted}{Should results be sorted (by input dimension and then alphabetically)}
}
\value{
See details
}
\description{
Returns information about the functions found in this package.
}
\details{
If fname is specified, this function returns a list with the number of input dimensions and a p x 2 matrix of input ranges. If fname is not specified, then `quackquack` returns a list of function names which satisfy the requirements specified by the other inputs. If no arguments are specified, then a list of all functions is returned.
}
\examples{
quack("borehole")

quack(input_dim=c(1, 2), input_cat = FALSE, response_type = "uni")
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.
}
