% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbits.R
\name{rabbits}
\alias{rabbits}
\title{The Rabbits Function (Logistic Growth)}
\usage{
rabbits(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 3. x[1]=P0 (initial pop), x[2]=t (time), x[3]=r (growth rate).}

\item{scale01}{When TRUE, all inputs are expected to be on the unit interval.}
}
\value{
Population size at time t.
}
\description{
Dimensions 3. Classic logistic growth with fixed carrying capacity K=1.
}
\details{
Logistic growth model:
\deqn{P(t) = \frac{P_0 e^{r t}}{1 + P_0 (e^{r t} - 1)}}
with P0 in [0,1], t in [0,1], r in [0.5,3].
}
\examples{
n <- 10
x <- matrix(runif(3*n), nrow=n)
y <- apply(x, 1, rabbits)
}
