% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripples.R
\name{ripples}
\alias{ripples}
\title{Ripples Function}
\usage{
ripples(x, scale01 = FALSE, fixed_coeff = TRUE, input_dims = 2, seed = NULL)
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{(No effect here). When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{fixed_coeff}{logical. Should coefficients be fixed at values from Rumsey et al (2022)? Ignored when \code{input_dims != 2}.}

\item{input_dims}{desired input dimension.}

\item{seed}{Optional seed for coefficient generation.}
}
\value{
Univariate response representing small scale structure
}
\description{
Dimensions 2. The ripples function was used by Rumsey et al (2022) to construct small scale structures for synthetic data mimicking the E3SM climate model.
}
\details{
The number of inputs can be made arbitrary, but only by setting \code{fixed_coeff = FALSE}. When \code{fixed_coeff = TRUE}, the ripples function exactly matches the function from Rumsey et al (2022). Otherwise, the coefficients are randomly generated from a standard normal distribution. All inputs are expected between zero and one
}
\examples{
n <- 10
x <- matrix(runif(2*n), nrow=n)
y <- apply(x, 1, ripples)
}
\references{
Rumsey et al reference needed (SPIE conference)
}
