% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot.R
\name{robot}
\alias{robot}
\title{Robot Arm Function}
\usage{
robot(x, scale01 = FALSE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
The response is the distance from the end of the robot arm to the origin, on the (u, v)-plane.
}
\description{
Dimensions: 8 Used commonly in neural network papers, models the position of a robot arm which has four segments. While the shoulder is fixed at the origin, the four segments each have length Li, and are positioned at an angle θi (with respect to the horizontal axis), for i = 1, …, 4
}
\details{
For details on this function, see \href{https://www.sfu.ca/~ssurjano/robot.html}{the VLSE}.
Parameter ranges:
\describe{
  \item{x1 = theta1}{angle of 1st arm segment, [0, 2*pi]}
  \item{x2 = theta2}{angle of 2nd arm segment, [0, 2*pi]}
  \item{x3 = theta3}{angle of 3rd arm segment, [0, 2*pi]}
  \item{x4 = theta4}{angle of 4th arm segment, [0, 2*pi]}
  \item{x5 = L1}{length of 1st arm segment,    [0, 1]}
  \item{x6 = L2}{length of 2nd arm segment,    [0, 1]}
  \item{x7 = L3}{length of 3rd arm segment,    [0, 1]}
  \item{x8 = L4}{length of 4th arm segment,    [0, 1]}
}
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, robot, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Ben-Ari, E. N., & Steinberg, D. M. (2007). Modeling data from computer experiments: an empirical comparison of kriging with MARS and projection pursuit regression. Quality Engineering, 19(4), 327-338.
}
