% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_study.R
\name{run_one_sim_case}
\alias{run_one_sim_case}
\title{==================================================
Fit models
==================================================}
\usage{
run_one_sim_case(
  rr,
  seed,
  fn,
  fnum,
  p,
  n,
  nsr,
  dsgn,
  n_test,
  conf_level,
  interval,
  method_names,
  my_fit,
  my_pred,
  verbose
)
}
\description{
==================================================
Fit models
==================================================
}
