% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_det_sharkfin.R
\name{sharkfin}
\alias{sharkfin}
\alias{sharkfin12}
\title{Sharkfin Function}
\usage{
sharkfin(x, scale01 = TRUE, steps = 3, active_dim = length(x))

sharkfin12(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{steps}{Default is 3. For larger values of \code{steps}, the function becomes more fractal-like and takes longer to run.}

\item{active_dim}{Only the first \code{min(active_dim, length(x))} variables will be used. All other variables are inert.}
}
\value{
Univariate response representing small scale structure
}
\description{
Dimensions default is 3, but can be any dimension. For inert variables, \code{active_dim} must be specified.
}
\details{
A non-smooth test function designed with trees in mind. As steps -> Inf, this function is nowhere-differentiable.
}
\examples{
fname <- "sharkfin"
n <- 10
p <- quack(fname, verbose=FALSE)$input_dim
x <- matrix(runif(n*p), nrow=n)
y <- eval_duq(fname, x)
}
\references{
Collins, G., Francom, D. and Rumsey, K., 2024. Bayesian projection pursuit regression. Statistics and Computing, 34(1), p.29.
}
