% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_column.R
\name{short_column}
\alias{short_column}
\title{Short Column}
\usage{
short_column(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 5. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval.}
}
\description{
Dimension 5.
}
\details{
The Short Column function models a short column with uncertain material properties and dimensions sbject to uncertain loads. Y is the yield stress, M is the bending moment, P is the axial force, b is the width of the cross section, and h is the depth of the cross section. Note that sometimes M and P are given a correlated prior (correlation coefficient = 0.5).
For more details on the Cantilever Beam, see \href{https://www.sfu.ca/~ssurjano/shortcol.html}{the VLSE}.
}
\examples{
X <- lhs::randomLHS(50, 7)
y <- apply(X, 1, short_column, scale01=TRUE)
}
\references{
Eldred, M. S., et al. "Investigation of reliability method formulations in DAKOTA/UQ." Structure and Infrastructure Engineering 3.3 (2007): 199-213.

Kuschel, Norbert, and Rüdiger Rackwitz. "Two basic problems in reliability-based structural optimization." Mathematical Methods of Operations Research 46 (1997): 309-333.
}
