% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_function_2d.R
\name{show_function_2d}
\alias{show_function_2d}
\title{The Foursquare Function}
\usage{
show_function_2d(fname, nx = 201, ncol = 100, pal = "terrain", dims = 1:2, ...)
}
\arguments{
\item{fname}{function name (from duqling)}

\item{nx}{grid size (one side)}

\item{ncol}{number of colors used in plot}

\item{pal}{palette passed to hcl.colors}

\item{dims}{length must be 2. Which dimensions should be plotted?}
}
\value{
A figure
}
\description{
Dimensions: 2
}
\details{
Plots 2 dimensions of a specified test functions
}
\examples{
show_function_2d("twin_galaxies")
show_function_2d("borehole", dims=c(1,6))
}
