% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_machine.R
\name{simple_machine}
\alias{simple_machine}
\title{Simple Machine}
\usage{
simple_machine(
  x,
  scale01 = FALSE,
  effort = seq(1, 10, by = 1),
  friction = 1/10,
  base_eff = 0
)
}
\arguments{
\item{x}{Inputs of dimension 1 to 3. Parameters are efficiency, friction and base-efficiency. If x has dimension less than 3, defaults are used for the other pars.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{effort}{Functional input. See details}

\item{friction}{friction value (only used if x is dim one or two)}

\item{base_eff}{base efficiency value (only used if x is dim one)}
}
\value{
The "output" of the simple machine with parameters x at effort levels specified by \code{effort}.
}
\description{
Dimensions 3. The simple machine computer model which is prominent in the Bayesian calibration literature. Functional response.
}
\details{
The simple machine is given by

\code{output = efficiency*effort/(1+effort*friction) + base efficiency}

Usual parameter ranges are given by efficiency in (0, 3), friction in (0, 50) and base efficiency in (-1, 1). See also \code{\link[duqling]{simple_machine_cm}}.
}
\examples{
n <- 10
x <- matrix(runif(3*n), nrow=n)
y <- apply(x, 1, simple_machine)
}
\references{
Brynjarsdóttir, Jenný, and Anthony OʼHagan. "Learning about physical parameters: The importance of model discrepancy." Inverse problems 30.11 (2014): 114007.

Rumsey, Kellin N. Methods of uncertainty quantification for physical parameters. Diss. The University of New Mexico, 2020.
}
