% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_machine.R
\name{simple_machine_cm}
\alias{simple_machine_cm}
\title{Simple Machine Computer Model (Piecewise Polynomial)}
\usage{
simple_machine_cm(x, scale01 = FALSE, effort = seq(1, 10, by = 1), order = 3)
}
\arguments{
\item{x}{Inputs of dimension 1 to 3. Parameters are efficiency, friction and base-efficiency. If x has dimension less than 3, defaults are used for the other pars.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{effort}{Functional input. See details}

\item{order}{Default 3. The dimension of x becomes 3 and 1 when \code{order = 2} and \code{order = 1} respectively. See Rumsey (2020) eq. 4.33 for details.}
}
\value{
The "output" of the simple machine with parameters x at effort levels specified by \code{effort}.
}
\description{
Dimensions 5. An attempt to model the simple machine while recognizing the presence of friction but using an incorrect mechanism. Setting order = 1 returns the linear model from Brynjarsdottir and O'Hagan.
}
\details{
A piecewise polyonmial approximation to the \code{\link[duqling]{simple_machine}}. Parameters consist of (alpha, gamma1, beta1, gamma2, beta2)
}
\examples{
n <- 10
x <- matrix(runif(3*n), nrow=n)
y <- apply(x, 1, simple_machine)
}
\references{
Brynjarsdóttir, Jenný, and Anthony OʼHagan. "Learning about physical parameters: The importance of model discrepancy." Inverse problems 30.11 (2014): 114007.

Rumsey, Kellin N. Methods of uncertainty quantification for physical parameters. Diss. The University of New Mexico, 2020.
}
