% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_det_simple_poly.R
\name{simple_poly}
\alias{simple_poly}
\title{Simple Polynomial Function from Rumsey et al 2023}
\usage{
simple_poly(x, beta = 1/9, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 2. See below for details.}

\item{beta}{default 1/9}

\item{scale01}{(Ignored here) When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
A scalar response
}
\description{
Dimensions: 2.
}
\details{
a simple polynomial function.
}
\examples{
fname <- "simple_poly"
n <- 10
p <- quack(fname, verbose=FALSE)$input_dim
x <- matrix(runif(n*p), nrow=n)
y <- eval_duq(fname, x)
}
\references{
Rumsey, K.N., Hardy, Z.K., Ahrens, C. and Vander Wiel, S., 2025. Co-active subspace methods for the joint analysis of adjacent computer models. Technometrics, 67(1), pp.133-146
}
