% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol.R
\name{sobol}
\alias{sobol}
\title{Sobol Function}
\usage{
sobol(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 3. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval.}
}
\description{
Sobol Function
}
\details{
Often used as an example able to compare sobol indices directly
}
\examples{
X <- lhs::randomLHS(50, 9)
y <- apply(X, 1, sobol, scale01=TRUE)
}
\references{
I.M. Sobol’, Sensitivity estimates for nonlinear mathematical models, Math. Model. Comput. Exp. 1 (1993) 407–414.
}
