% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_det_squiggle.R
\name{squiggle}
\alias{squiggle}
\title{The Squiggle Function}
\usage{
squiggle(x, scale01 = TRUE, sigma = 0.05)
}
\arguments{
\item{x}{Inputs of dimension (at least) 2. See below for details.}

\item{scale01}{(No effect) When TRUE, inputs are expected to be given on unit interval.}

\item{sigma}{A scaling parameter}
}
\description{
The Squiggle Function
}
\details{
A bivariate squiggle function which has regions of low activity as well as low activity regions.
When noise != 0, the function is stochastic with a heteroskedastic error function.
}
\examples{
fname <- "squiggle"
n <- 10
p <- quack(fname, verbose=FALSE)$input_dim
x <- matrix(runif(n*p), nrow=n)
y <- eval_duq(fname, x)

show_function_2d("squiggle")
}
\references{
Add reference here, if we ever publish with this function.
}
