% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{star2}
\alias{star2}
\title{2D Star Jump Process}
\usage{
star2(x, scale01 = TRUE, mu = c(1, -1), sigma = c(3, 3), ell = c(0.1, 0.2))
}
\arguments{
\item{x}{Inputs of dimension (at least) 2.}

\item{scale01}{Ignored.}

\item{mu}{mean of each process.}

\item{sigma}{standard deviation of each process.}

\item{ell}{lengthscale of RBF kernels for each process.}
}
\value{
Scalar output
}
\description{
Dimensions: 2. A simple star jump process
}
\details{
Uses a fixed mask in the shape of a star to determine which of two smooth processes will be returned at each location. See reference for details, but note that this implementation uses a determinstic error structure based on a linear combination of 5 RBF kernels for each process.
}
\examples{
n <- 2
x <- matrix(runif(50), nrow=n)
y <- apply(x, 1, star2)
}
\references{
Flowers, A. R., Franck, C. T., Binois, M., Park, C., & Gramacy, R. B. (2025). Modular Jump Gaussian Processes. arXiv preprint arXiv:2505.15557.
}
