% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steel_column.R
\name{steel_column}
\alias{steel_column}
\title{Steel Column}
\usage{
steel_column(x, scale01 = TRUE, b = 300, d = 20, h = 300)
}
\arguments{
\item{x}{Inputs of dimension (at least) 5. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval.}

\item{b}{default parameter}

\item{d}{default parameter}

\item{h}{default parameter}
}
\description{
Dimension 9. Models the "reliability" of a steel column.
}
\details{
This function is often used to explore the trade-off between cost and reliability of a steel column. The cost is \code{b*d + 5*h} which are parameters giving the mean flange breadth, flange thickness, and profile height, respectively. Non-normal distributions are often used for the priors.
For details, see \href{https://www.sfu.ca/~ssurjano/steelcol.html}{the VLSE}.
}
\examples{
X <- lhs::randomLHS(50, 7)
y <- apply(X, 1, steel_column, scale01=TRUE)
}
\references{
Kuschel, Norbert, and Rüdiger Rackwitz. "Two basic problems in reliability-based structural optimization." Mathematical Methods of Operations Research 46 (1997): 309-333.
}
