% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piston.R
\name{stochastic_piston}
\alias{stochastic_piston}
\title{Stochastic Piston Simulation Function}
\usage{
stochastic_piston(
  x,
  scale01 = TRUE,
  Ta_generate = function() rbeta(1, 10, 15),
  P0_generate = function() runif(1, 0.49, 0.51)
)
}
\arguments{
\item{x}{Inputs of dimension (at least) 5. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{Ta_generate}{The generating distribution for the ambient temperature (on (0,1) scale). Default is `rbeta(1, 10, 1.5)`}

\item{P0_generate}{The generating distribution for the atmospheric pressure (on (0,1) scale). Default is `runif(1, 0.49, 0.51)`}
}
\value{
The response C is cycle time (the time it takes to complete one cycle), in seconds
}
\description{
Dimensions: 5 The Piston Simulation function models the circular motion of a piston within a cylinder. It involves a chain of nonlinear functions.
}
\details{
For details on the piston function, see \href{https://www.sfu.ca/~ssurjano/piston.html}{the VLSE}.
Parameter ranges:
\describe{
  \item{x1 = M}{piston weight (kg), [30, 60]}
  \item{x2 = S}{piston surface area (m^2), [0.020, 0.045]}
  \item{x3 = V0}{initial gas volume (m^3), [0.002, 0.010]}
  \item{x4 = k}{spring coefficient (N/m), [1000, 5000]}
  \item{x5 = T0}{filling gas temperature (K), [340, 360]}
}
This is a stochastic version of the piston function. Note that the range of the piston surface area is different from the canonical piston function. The distribution of the response is heavily influenced by the stochastic process governing ambient temperature. The response distribution is less homoskedastic for larger ranges of atmospheric pressure.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, piston, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Kenett, R., & Zacks, S. (1998). Modern industrial statistics: design and control of quality and reliability. Pacific Grove, CA: Duxbury press.
}
