% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sulfur.R
\name{sulfur}
\alias{sulfur}
\title{Sulfur}
\usage{
sulfur(x, scale01 = TRUE, b = 300, d = 20, h = 300)
}
\arguments{
\item{x}{Inputs of dimension (at least) 5. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval.}
}
\description{
Dimension 9. Models the radiative forcing of sulfur with uncertain inputs. High degree of interaction (but additive in log space).
}
\details{
This function models the radiative forcing of sulfur by sulfate aerosols, in W/m^2. Note that all priors are lognormal.
For details, see \href{https://www.sfu.ca/~ssurjano/sulf.html}{the VLSE}. NOTE: There is a problem with dimensional analysis of the Tatang paper. There is a hidden (or missing) W in the denominator. As such, the results do not match the paper perfectly. Use this function at your own risk.
}
\examples{
X <- lhs::randomLHS(50, 7)
y <- apply(X, 1, sulfur, scale01=TRUE)
}
\references{
Tatang, Menner A., et al. "An efficient method for parametric uncertainty analysis of numerical geophysical models." Journal of Geophysical Research: Atmospheres 102.D18 (1997): 21925-21932

Charlson, Robert J., et al. "Climate forcing by anthropogenic aerosols." Science 255.5043 (1992): 423-430.
}
