% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_scalar_det_turbulence.R
\name{turb12}
\alias{turb12}
\title{12D Turbulence-Inspired Toy Function}
\usage{
turb12(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension at least 12.}

\item{scale01}{Are inputs given on (0, 1) scale?}
}
\value{
Scalar output.
}
\description{
A deterministic 12-dimensional toy model inspired by simple algebraic
eddy-viscosity ideas. This is not a true RANS closure model, but is meant
to mimic some of the nonlinear structure seen in turbulence surrogates:
wall damping, strain dependence, Reynolds-number effects, and dissipation.
}
\details{
Inputs are interpreted as:
\describe{
  \item{x1}{log-Reynolds-number surrogate}
  \item{x2}{wall-distance surrogate}
  \item{x3}{strain-rate magnitude surrogate}
  \item{x4}{mixing length surrogate}
  \item{x5}{eddy-viscosity coefficient}
  \item{x6}{wall-damping coefficient}
  \item{x7}{pressure-gradient surrogate}
  \item{x8}{curvature/separation surrogate}
  \item{x9}{production coefficient}
  \item{x10}{dissipation coefficient}
  \item{x11}{anisotropy surrogate}
  \item{x12}{large-scale shear surrogate}
}
}
\examples{
fname <- "turb12"
n <- 10
p <- quack(fname, verbose=FALSE)$input_dim
x <- matrix(runif(n * p), nrow = n)
y <- eval_duq(fname, x)
}
