% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin_galaxies.R
\name{twin_galaxies}
\alias{twin_galaxies}
\title{Synthetic Twin Galaxy Function}
\usage{
twin_galaxies(x, scale01 = FALSE, fixed_coeff = TRUE, seed = NULL)
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{fixed_coeff}{Logical. See \code{duqling::ripples} for details.}

\item{seed}{Optional. See \code{duqling::ripples} for details.}
}
\value{
Univariate response representing small scale structure
}
\description{
Dimensions 2. A synthetic computer model from Rumsey et al (2022) which was designed to mimic the challenges and features of the E3SM climate model. The name is derived purely from the appearance of the computer model when plotting.
}
\details{
Function with long range and small scale structures and prominent continent-esque features (the twin galaxies). See Rumsey et al (2022) for details along with the \code{duqling::lim_polyonmial}, \code{duqling::grlee2} and \code{duqling::ripples} functions. Inputs are called lon and lat with ranges (0, 360) and (-90, 90) respectively, due to the originiation of this function with climate models.
}
\examples{
n <- 10
x <- matrix(runif(2*n), nrow=n)
y <- apply(x, 1, twin_galaxies)
}
\references{
Rumsey, Kellin, et al. "A hierarchical sparse Gaussian process for in situ inference in expensive physics simulations." Applications of machine learning 2022. Vol. 12227. SPIE, 2022.
}
