% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana.R
\name{banana}
\alias{banana}
\title{Rosenbrock's Banana Function}
\usage{
banana(x, scale01 = TRUE, ab = c(1, 100))
}
\arguments{
\item{x}{Inputs of dimension (at least) 2. See below for details.}

\item{scale01}{(No effect) When TRUE, inputs are expected to be given on unit interval.}

\item{ab}{Hyperparameters to Rosenbrocks banana}
}
\description{
Rosenbrock's Banana Function
}
\details{
A non-convec function introduced by Howard H. Rosenbrock in 1960. It is a difficult optimization test-problem, with a unique minimum at (1,1).
}
\examples{
n1 <- 100 #Number of observations in x1
n2 <- 70  #Number of observations in x2
x1grid <- seq(0, 1, length.out=100)
x2grid <- seq(0, 1, length.out=70)
X <- expand.grid(x1grid, x2grid)
y <- apply(X, 1, banana, scale01=TRUE)
image(matrix(y, nrow=length(X)), zlim=c(-1,3.5))
}
\references{
Rosenbrock, HoHo. "An automatic method for finding the greatest or least value of a function." The computer journal 3.3 (1960): 175-184.
}
