% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_UQ_data.R
\name{data_quack}
\alias{data_quack}
\title{Get UQ data for emulation}
\usage{
data_quack(
  raw = FALSE,
  dname = NULL,
  input_dim = NULL,
  input_cat_dim = NULL,
  n = NULL,
  output_dim = NULL
)
}
\arguments{
\item{raw}{Logical. When TRUE, full dataset information is given. Otherwise, returns list of scalar-response datasets for emulation.}

\item{dname}{Vector of function names to filter.}

\item{input_dim}{Vector of input dimensions (quantitative only) to filter.}

\item{input_cat_dim}{Vector of input dimensions (categorical only) to filter.}

\item{n}{Vector of sample sizes to filter.}

\item{output_dim}{Vector of output dimensions to filter.}
}
\value{
A table with data set names meeting the filter requirements
}
\description{
Load in UQ datasets from the UQdataverse (harvard)
}
\details{
Wrapper to read in various UQ datasets (from harvard dataverse)
}
\examples{
tab <- data_quack(raw=TRUE)
get_UQ_data(tab$dname[1])

tab <- data_quack(raw=FALSE)
get_emulation_data(tab$dname[1])
}
