% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detpep_curve.R
\name{detpep_curve}
\alias{detpep_curve}
\title{The Dette & Pep Curved Function}
\usage{
detpep_curve(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 3. See below for details.}

\item{scale01}{(No effect) When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
Response.
}
\description{
Dimensions 3.
}
\details{
This highly curved function is used for the comparison of computer experiment designs.
}
\examples{
n <- 10
x <- matrix(runif(3*n), nrow=n)
y <- apply(x, 1, detpep_curve, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Dette, Holger, and Andrey Pepelyshev. "Generalized Latin hypercube design for computer experiments." Technometrics 52.4 (2010): 421-429.
}
