% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duq.R
\name{duq}
\alias{duq}
\title{duq}
\usage{
duq(x, f, ...)
}
\arguments{
\item{x}{An nxp matrix of inputs.}

\item{f}{A function, usually from the duqling package}

\item{...}{additional functions to be passed to f}
}
\value{
See details
}
\description{
Call functions from duqling namespace
}
\details{
If fname is specified, this function returns a list with the number of input dimensions and a p x 2 matrix of input ranges. If fname is not specified, then `quackquack` returns a list of function names which satisfy the requirements specified by the other inputs. If no arguments are specified, then a list of all functions is returned.
}
\examples{
n <- 100
p <- 10
X <- lhs::randomLHS(n, p)
duq(X, borehole)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.
}
