% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_sim_study.R
\name{heatmap_sim_study}
\alias{heatmap_sim_study}
\title{Heatmap of Method Rankings Across Simulation Scenarios}
\usage{
heatmap_sim_study(
  df,
  metric = "CRPS",
  methods = NULL,
  group_by = "fname",
  ties_method = "min",
  show_values = NULL,
  color_scale = "turbo",
  orientation = "vertical",
  path = NULL
)
}
\arguments{
\item{df}{A data frame from \code{run_sim_study()}.}

\item{metric}{Character. The performance metric to plot (e.g., \code{"CRPS"}, \code{"RMSE"}). By default, ranks are shown; use \code{"CRPS_raw"} or similar to show raw values.}

\item{methods}{Optional character vector of methods to include.}

\item{group_by}{Character vector of columns that define each simulation scenario (default = "fname").}

\item{ties_method}{How to handle ties in ranking. Passed to \code{rank()}. Default = "min".}

\item{show_values}{Logical. Should text given the average "metric" value (rounded to 2 decimal places) be added to the plot?}

\item{color_scale}{Either a string specifying one of the 8 viridis color map options
("magma", "inferno", "plasma", "viridis", "cividis", "rocket", "mako", "turbo"),
or a character vector of two colors (e.g., \code{c("white", "black")}) to be used
as a custom gradient from low to high rank. Default is \code{"turbo"}.}

\item{path}{Optional file path to save plot (e.g., "figs/CRPS_results.png"). If NULL, the plot is returned instead of saved.}
}
\value{
A ggplot heatmap.
}
\description{
Heatmap of Method Rankings Across Simulation Scenarios
}
\examples{
\dontrun{
heatmap_sim_study(df, metric = "CRPS", group_by = c("fname", "NSR"))
}
}
