% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ishigami.R
\name{ishigami}
\alias{ishigami}
\title{Ishigami Function}
\usage{
ishigami(x, scale01 = TRUE, ab = c(7, 0.1))
}
\arguments{
\item{x}{Inputs of dimension (at least) 3. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval.}

\item{ab}{Hyperparameters to the Ishigami function}
}
\description{
Ishigami Function
}
\details{
Often used as an example, because it exhibits strong nonlinearity and nonmonotonicity. It has a peculiar dependence on x3. By default, the distribution of the inputs are x_i ~ Uniform(-pi, pi)
}
\examples{
X <- lhs::randomLHS(50, 3)
y <- apply(X, 1, ishigami, scale01=TRUE)
}
\references{
Ishigami, T., & Homma, T. (1990 December). An importance quantification technique in uncertainty analysis for computer models. In Uncertainty Modeling and Analysis, 1990.
}
