% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoplot_sim_study.R
\name{paretoplot_sim_study}
\alias{paretoplot_sim_study}
\title{Pareto Plot of Accuracy vs Time}
\usage{
paretoplot_sim_study(
  df,
  metric = c("CRPS", "t_tot"),
  methods = NULL,
  highlight_method = NA,
  path = NULL
)
}
\arguments{
\item{df}{Data frame from \code{run_sim_study()}.}

\item{metric}{Length-2 character vector. The x and y metrics to compare. Default is \code{c("CRPS", "t_tot")}.}

\item{methods}{Optional character vector of methods to include.}

\item{highlight_method}{Character vector of methods to label and color. Use \code{NA} to color all (default), or \code{NULL} to color none.}

\item{path}{Optional file path to save plot (e.g., "figs/CRPS_results.png"). If NULL, the plot is returned instead of saved.}
}
\value{
A ggplot object.
}
\description{
Plot tradeoff between performance and runtime for each method.
}
\examples{
\dontrun{
paretoplot_sim_study(results)
paretoplot_sim_study(results, highlight_method = c("GP", "BART"))
paretoplot_sim_study(results, metric = c("RMSE", "t_tot"), highlight_method = NULL)
}
}
