% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quackquack.R
\name{quack}
\alias{quack}
\title{Quack (Gives Information about functions in this package)}
\usage{
quack(
  fname = NULL,
  input_dim = NULL,
  response = NULL,
  stochastic = NULL,
  has_categorical = NULL,
  sorted = TRUE,
  show_sigma = FALSE,
  ...
)
}
\arguments{
\item{fname}{A string, the name of a function in this package}

\item{input_dim}{a vector specifying the input dimension of the function}

\item{response}{a string in the set c("all", "univariate", "multivariate", "functional") specifying which response type is requested.}

\item{stochastic}{logical. Is function response stochastic?}

\item{has_categorical}{logical, should functions with categorical inputs be included?}

\item{sorted}{Should results be sorted (by input dimension and then alphabetically)}

\item{show_sigma}{Should we try to retrieve the variance of the function response across its default input range? Not available for all functions. (Not actively supported, use with caution).}
}
\value{
See details
}
\description{
Returns information about the functions found in this package.
}
\details{
If fname is specified, this function returns a list with the number of input dimensions and a p x 2 matrix of input ranges. If fname is not specified, then `quackquack` returns a list of function names which satisfy the requirements specified by the other inputs. If no arguments are specified, then a list of all functions is returned.
}
\examples{
quack("borehole")

quack(input_dim=c(1, 2), has_categorical = FALSE, response = "univariate")
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.
}
