% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkfin.R
\name{sharkfin}
\alias{sharkfin}
\title{Sharkfin Function}
\usage{
sharkfin(x, scale01 = TRUE, steps = 3, active_dim = length(x))
}
\arguments{
\item{x}{Inputs of dimension (at least) 8. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}

\item{steps}{Default is 3. For larger values of \code{steps}, the function becomes more fractal-like and takes longer to run.}

\item{active_dim}{Only the first \code{min(active_dim, length(x))} variables will be used. All other variables are inert.}
}
\value{
Univariate response representing small scale structure
}
\description{
Dimensions default is 3, but can be any dimension. For inert variables, \code{active_dim} must be specified.
}
\details{
A non-smooth test function designed with trees in mind. As steps -> Inf, this function is nowhere-differentiable.
}
\examples{
n <- 10
x <- matrix(runif(3*n), nrow=n)
y <- apply(x, 1, sharkfin)
}
\references{
Collins, Gavin, Devin Francom, and Kellin Rumsey. "Bayesian projection pursuit regression." arXiv preprint arXiv:2210.09181 (2022).
}
