% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_poly.R
\name{simple_poly}
\alias{simple_poly}
\title{Simple Polynomial Function from Rumsey et al 2023}
\usage{
simple_poly(x, beta = 1/9, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 2. See below for details.}

\item{beta}{default 1/9}

\item{scale01}{(Ignored here) When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
A scalar response
}
\description{
Dimensions: 2.
}
\details{
a simple polynomial function.
}
\examples{
n <- 10
x <- matrix(runif(8*n), nrow=n)
y <- apply(x, 1, simple_poly, scale01=TRUE)
}
\references{
Rumsey, Kellin N., et al. "Co-Active Subspace Methods for the Joint Analysis of Adjacent Computer Models." arXiv preprint arXiv:2311.18146 (2023).
}
