% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squiggle.R
\name{squiggle}
\alias{squiggle}
\title{The Squiggle Function}
\usage{
squiggle(x, scale01 = TRUE, noise = 0.05)
}
\arguments{
\item{x}{Inputs of dimension (at least) 2. See below for details.}

\item{scale01}{(No effect) When TRUE, inputs are expected to be given on unit interval.}

\item{sigma}{A scaling parameter}
}
\description{
The Squiggle Function
}
\details{
A bivariate squiggle function which has regions of low activity as well as low activity regions.
When noise != 0, the function is stochastic with a heteroskedastic error function.
}
\examples{
n1 <- 100 #Number of observations in x1
n2 <- 70  #Number of observations in x2
x1grid <- seq(0, 1, length.out=100)
x2grid <- seq(0, 1, length.out=70)
X <- expand.grid(x1grid, x2grid)
y <- apply(X, 1, squiggle)
image(matrix(y, nrow=length(X)), zlim=c(-1,3.5))
}
\references{
Add reference here, if we ever publish with this function.
}
