% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detpep_curve.R
\name{welch20}
\alias{welch20}
\title{The Welch Screening Function}
\usage{
welch20(x, scale01 = TRUE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 20. See below for details.}

\item{scale01}{(No effect) When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
Response.
}
\description{
Dimensions 20.
}
\details{
A challenging high-dimensional problem used for variable screening purposes
}
\examples{
n <- 10
x <- matrix(runif(20*n), nrow=n)
y <- apply(x, 1, welch20, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Welch, W. J., Buck, R. J., Sacks, J., Wynn, H. P., Mitchell, T. J., & Morris, M. D. (1992). Screening, predicting, and computer experiments. Technometrics, 34(1), 15-25.
}
