% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingweight.R
\name{wingweight}
\alias{wingweight}
\title{Wing Weight Function}
\usage{
wingweight(x, scale01 = FALSE)
}
\arguments{
\item{x}{Inputs of dimension (at least) 10. See below for details.}

\item{scale01}{When TRUE, inputs are expected to be given on unit interval and are internally adjusted to their native range. Default is FALSE}
}
\value{
The response is the wing's wieght
}
\description{
Dimensions: 10 Models a light aircraft wing.
}
\details{
For details on this function, see \href{https://www.sfu.ca/~ssurjano/robot.html}{the VLSE}.
For parameter ranges, see the \href{https://www.sfu.ca/~ssurjano/robot.html}{the VLSE}.
}
\examples{
n <- 10
x <- matrix(runif(10*n), nrow=n)
y <- apply(x, 1, wingweight, scale01=TRUE)
}
\references{
Surjanovic, Sonja, and Derek Bingham. "Virtual library of simulation experiments: test functions and datasets." Simon Fraser University, Burnaby, BC, Canada, accessed May 13 (2013): 2015.

Forrester, A., Sobester, A., & Keane, A. (2008). Engineering design via surrogate modelling: a practical guide. Wiley.

Moon, H. (2010). Design and Analysis of Computer Experiments for Screening Input Variables (Doctoral dissertation, Ohio State University).
}
